requirements.txt
psycopg2-binary
pandas
scikit-learn
openai
transformers
langchain
